CMS-3014 Turn PageTemplateCommands into api.content.page.PageTemplateService
