+++ modules/wem-admin/src/main/java/com/enonic/wem/admin/rest/resource/content/page/UpdatePageJson.java
@@ -11,11 +11,11 @@ import com.enonic.wem.admin.json.content.page.region.PageRegionsJson;
 import com.enonic.wem.admin.json.content.page.region.RegionJson;
 import com.enonic.wem.admin.json.data.DataJson;
 import com.enonic.wem.admin.json.data.RootDataSetJson;
-import com.enonic.wem.api.command.content.page.UpdatePage;
 import com.enonic.wem.api.content.ContentId;
 import com.enonic.wem.api.content.page.Page;
 import com.enonic.wem.api.content.page.PageEditor;
 import com.enonic.wem.api.content.page.PageTemplateKey;
+import com.enonic.wem.api.content.page.UpdatePage;
 
 import static com.enonic.wem.api.content.page.Page.editPage;
