+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/CreatePage.java
@@ -1,11 +1,9 @@
-package com.enonic.wem.api.command.content.page;
+package com.enonic.wem.api.content.page;
 
 
 import com.enonic.wem.api.command.Command;
 import com.enonic.wem.api.content.Content;
 import com.enonic.wem.api.content.ContentId;
-import com.enonic.wem.api.content.page.PageTemplateKey;
-import com.enonic.wem.api.content.page.PageRegions;
 import com.enonic.wem.api.data.RootDataSet;
 
 public class CreatePage
