+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/CreatePageDescriptor.java
@@ -1,19 +1,16 @@
-package com.enonic.wem.api.command.content.page.layout;
+package com.enonic.wem.api.content.page;
 
 
 import com.google.common.base.Preconditions;
 
 import com.enonic.wem.api.command.Command;
-import com.enonic.wem.api.content.page.ComponentDescriptorName;
-import com.enonic.wem.api.content.page.layout.LayoutDescriptor;
-import com.enonic.wem.api.content.page.layout.LayoutDescriptorKey;
 import com.enonic.wem.api.content.page.region.RegionDescriptors;
 import com.enonic.wem.api.form.Form;
 
-public class CreateLayoutDescriptor
-    extends Command<LayoutDescriptor>
+public class CreatePageDescriptor
+    extends Command<PageDescriptor>
 {
-    private LayoutDescriptorKey key;
+    private PageDescriptorKey key;
 
     private ComponentDescriptorName name;
 
@@ -23,41 +20,47 @@ public class CreateLayoutDescriptor
 
     private RegionDescriptors regions;
 
-    public CreateLayoutDescriptor()
+    public CreatePageDescriptor()
     {
     }
 
-    public CreateLayoutDescriptor key( final LayoutDescriptorKey key )
+    public CreatePageDescriptor key( final PageDescriptorKey key )
     {
         this.key = key;
         return this;
     }
 
-    public CreateLayoutDescriptor name( final ComponentDescriptorName name )
+    public CreatePageDescriptor name( final ComponentDescriptorName name )
     {
         this.name = name;
         return this;
     }
 
-    public CreateLayoutDescriptor displayName( final String displayName )
+    public CreatePageDescriptor name( final String name )
+    {
+        this.name = new ComponentDescriptorName( name );
+        return this;
+    }
+
+    public CreatePageDescriptor displayName( final String displayName )
     {
         this.displayName = displayName;
         return this;
     }
 
-    public CreateLayoutDescriptor config( final Form config )
+    public CreatePageDescriptor config( final Form config )
     {
         this.config = config;
         return this;
     }
 
-    public CreateLayoutDescriptor regions( final RegionDescriptors value )
+    public CreatePageDescriptor regions( final RegionDescriptors value )
     {
         this.regions = value;
         return this;
     }
 
-    public LayoutDescriptorKey getKey()
+    public PageDescriptorKey getKey()
     {
         return key;
     }
@@ -88,7 +91,7 @@ public class CreateLayoutDescriptor
         Preconditions.checkNotNull( key, "key is required" );
         Preconditions.checkNotNull( name, "name is required" );
         Preconditions.checkNotNull( displayName, "displayName is required" );
-        Preconditions.checkNotNull( config, "config cannot be null" );
-        Preconditions.checkNotNull( regions, "regions cannot be null" );
+        Preconditions.checkNotNull( regions, "regions are required" );
+        Preconditions.checkNotNull( config, "config is required" );
     }
 }
