+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/GetPageTemplateByKey.java
@@ -1,38 +1,27 @@
-package com.enonic.wem.api.command.content.page;
+package com.enonic.wem.api.content.page;
 
 
 import com.google.common.base.Preconditions;
 
 import com.enonic.wem.api.command.Command;
-import com.enonic.wem.api.content.page.PageTemplate;
-import com.enonic.wem.api.content.page.PageTemplateEditor;
-import com.enonic.wem.api.content.page.PageTemplateKey;
 import com.enonic.wem.api.content.site.SiteTemplateKey;
 
-public final class UpdatePageTemplate
+public class GetPageTemplateByKey
     extends Command<PageTemplate>
 {
     private SiteTemplateKey siteTemplateKey;
 
     private PageTemplateKey key;
 
-    private PageTemplateEditor editor;
-
-    public UpdatePageTemplate siteTemplateKey( final SiteTemplateKey siteTemplateKey )
+    public GetPageTemplateByKey siteTemplateKey( final SiteTemplateKey siteTemplateKey )
     {
         this.siteTemplateKey = siteTemplateKey;
         return this;
     }
 
-    public UpdatePageTemplate key( final PageTemplateKey key )
+    public GetPageTemplateByKey key( final PageTemplateKey key )
     {
         this.key = key;
-        return this;
-    }
-
-    public UpdatePageTemplate editor( final PageTemplateEditor editor )
-    {
-        this.editor = editor;
         return this;
     }
 
@@ -51,10 +40,5 @@ public final class UpdatePageTemplate
     public PageTemplateKey getKey()
     {
         return key;
-    }
-
-    public PageTemplateEditor getEditor()
-    {
-        return editor;
     }
 }
