+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/PageCommands.java
@@ -1,4 +1,4 @@
-package com.enonic.wem.api.command.content.page;
+package com.enonic.wem.api.content.page;
 
 
 public final class PageCommands
