+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/PageDescriptorService.java
@@ -1,7 +1,5 @@
 package com.enonic.wem.api.content.page;
 
-import com.enonic.wem.api.command.content.page.CreatePageDescriptor;
-
 public interface PageDescriptorService
 {
     PageDescriptor getByKey( final PageDescriptorKey key );
