+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/UpdatePage.java
@@ -1,10 +1,9 @@
-package com.enonic.wem.api.command.content.page;
+package com.enonic.wem.api.content.page;
 
 
 import com.enonic.wem.api.command.Command;
 import com.enonic.wem.api.content.Content;
 import com.enonic.wem.api.content.ContentId;
-import com.enonic.wem.api.content.page.PageEditor;
 
 public class UpdatePage
     extends Command<Content>
