+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/UpdatePageTemplate.java
@@ -1,29 +1,35 @@
-package com.enonic.wem.api.command.content.page;
+package com.enonic.wem.api.content.page;
 
 
 import com.google.common.base.Preconditions;
 
 import com.enonic.wem.api.command.Command;
-import com.enonic.wem.api.content.page.PageTemplate;
-import com.enonic.wem.api.content.page.PageTemplateKey;
 import com.enonic.wem.api.content.site.SiteTemplateKey;
 
-public class GetPageTemplateByKey
+public final class UpdatePageTemplate
     extends Command<PageTemplate>
 {
     private SiteTemplateKey siteTemplateKey;
 
     private PageTemplateKey key;
 
-    public GetPageTemplateByKey siteTemplateKey( final SiteTemplateKey siteTemplateKey )
+    private PageTemplateEditor editor;
+
+    public UpdatePageTemplate siteTemplateKey( final SiteTemplateKey siteTemplateKey )
     {
         this.siteTemplateKey = siteTemplateKey;
         return this;
     }
 
-    public GetPageTemplateByKey key( final PageTemplateKey key )
+    public UpdatePageTemplate key( final PageTemplateKey key )
     {
         this.key = key;
+        return this;
+    }
+
+    public UpdatePageTemplate editor( final PageTemplateEditor editor )
+    {
+        this.editor = editor;
         return this;
     }
 
@@ -42,5 +48,10 @@ public class GetPageTemplateByKey
     public PageTemplateKey getKey()
     {
         return key;
+    }
+
+    public PageTemplateEditor getEditor()
+    {
+        return editor;
     }
 }
