+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/layout/GetLayoutDescriptorsByModules.java
@@ -1,18 +1,17 @@
-package com.enonic.wem.api.command.content.page.part;
+package com.enonic.wem.api.content.page.layout;
 
 
 import com.google.common.base.Preconditions;
 
 import com.enonic.wem.api.command.Command;
-import com.enonic.wem.api.content.page.part.PartDescriptors;
 import com.enonic.wem.api.module.ModuleKeys;
 
-public class GetPartDescriptorsByModules
-    extends Command<PartDescriptors>
+public class GetLayoutDescriptorsByModules
+    extends Command<LayoutDescriptors>
 {
     private final ModuleKeys moduleKeys;
 
-    public GetPartDescriptorsByModules( final ModuleKeys moduleKeys )
+    public GetLayoutDescriptorsByModules( final ModuleKeys moduleKeys )
     {
         this.moduleKeys = moduleKeys;
     }
