+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/part/CreatePartDescriptor.java
@@ -1,12 +1,10 @@
-package com.enonic.wem.api.command.content.page.part;
+package com.enonic.wem.api.content.page.part;
 
 
 import com.google.common.base.Preconditions;
 
 import com.enonic.wem.api.command.Command;
 import com.enonic.wem.api.content.page.ComponentDescriptorName;
-import com.enonic.wem.api.content.page.part.PartDescriptor;
-import com.enonic.wem.api.content.page.part.PartDescriptorKey;
 import com.enonic.wem.api.form.Form;
 
 public class CreatePartDescriptor
