+++ modules/wem-api/src/main/java/com/enonic/wem/api/content/page/part/PartDescriptorService.java
@@ -1,6 +1,5 @@
 package com.enonic.wem.api.content.page.part;
 
-import com.enonic.wem.api.command.content.page.part.CreatePartDescriptor;
 import com.enonic.wem.api.module.ModuleKeys;
 
 public interface PartDescriptorService
