+++ modules/wem-core/src/main/java/com/enonic/wem/core/content/page/CreatePageHandler.java
@@ -2,9 +2,9 @@ package com.enonic.wem.core.content.page;
 
 import com.enonic.wem.api.command.Commands;
 import com.enonic.wem.api.command.content.UpdateContent;
-import com.enonic.wem.api.command.content.page.CreatePage;
 import com.enonic.wem.api.content.Content;
 import com.enonic.wem.api.content.editor.ContentEditor;
+import com.enonic.wem.api.content.page.CreatePage;
 import com.enonic.wem.api.content.page.Page;
 import com.enonic.wem.core.command.CommandHandler;
