+++ modules/wem-core/src/main/java/com/enonic/wem/core/content/page/GetPageTemplatesBySiteTemplateHandler.java
@@ -2,8 +2,8 @@ package com.enonic.wem.core.content.page;
 
 
 import com.enonic.wem.api.command.Commands;
-import com.enonic.wem.api.command.content.page.GetPageTemplatesBySiteTemplate;
 import com.enonic.wem.api.command.content.site.GetSiteTemplateByKey;
+import com.enonic.wem.api.content.page.GetPageTemplatesBySiteTemplate;
 import com.enonic.wem.api.content.site.SiteTemplate;
 import com.enonic.wem.core.command.CommandHandler;
