+++ modules/wem-core/src/main/java/com/enonic/wem/core/content/page/UpdatePageHandler.java
@@ -2,12 +2,12 @@ package com.enonic.wem.core.content.page;
 
 import com.enonic.wem.api.command.Commands;
 import com.enonic.wem.api.command.content.UpdateContent;
-import com.enonic.wem.api.command.content.page.UpdatePage;
 import com.enonic.wem.api.content.Content;
 import com.enonic.wem.api.content.ContentNotFoundException;
 import com.enonic.wem.api.content.editor.ContentEditor;
 import com.enonic.wem.api.content.page.Page;
 import com.enonic.wem.api.content.page.PageNotFoundException;
+import com.enonic.wem.api.content.page.UpdatePage;
 import com.enonic.wem.core.command.CommandHandler;
 
 import static com.enonic.wem.api.content.Content.editContent;
